+++ app/client/cli/keys.go
@@ -4,13 +4,15 @@ import (
 	"bytes"
 	"encoding/hex"
 	"fmt"
+	"path/filepath"
+	"strconv"
+	"strings"
+
 	"github.com/pokt-network/pocket/logger"
 	"github.com/pokt-network/pocket/shared/codec"
-	"github.com/pokt-network/pocket/shared/converters"
+	coreTypes "github.com/pokt-network/pocket/shared/core/types"
 	"github.com/pokt-network/pocket/shared/crypto"
-	utilTypes "github.com/pokt-network/pocket/utility/types"
-	"path/filepath"
-	"strings"
+	"github.com/pokt-network/pocket/shared/utils"
 
 	"github.com/pokt-network/pocket/app/client/keybase"
 	"github.com/spf13/cobra"
@@ -23,6 +25,9 @@ var (
 	importAs   string
 	hint       string
 	newPwd     string
+	storeChild bool
+	childPwd   string
+	childHint  string
 )
 
 func init() {
@@ -45,6 +50,7 @@ func NewKeysCommand() *cobra.Command {
 	importCmds := keysImportCommands()
 	signMsgCmds := keysSignMsgCommands()
 	signTxCmds := keysSignTxCommands()
+	slipCmds := keysSlipCommands()
 
 	// Add --pwd and --hint flags
 	applySubcommandOptions(createCmds, attachPwdFlagToSubcommands())
@@ -77,6 +83,12 @@ func NewKeysCommand() *cobra.Command {
 	applySubcommandOptions(signTxCmds, attachInputFlagToSubcommands())
 	applySubcommandOptions(signTxCmds, attachOutputFlagToSubcommands())
 
+	// Add --pwd, --store_child, --child_pwd, --child_hint flags
+	applySubcommandOptions(slipCmds, attachPwdFlagToSubcommands())
+	applySubcommandOptions(slipCmds, attachStoreChildFlagToSubcommands())
+	applySubcommandOptions(slipCmds, attachChildPwdFlagToSubcommands())
+	applySubcommandOptions(slipCmds, attachChildHintFlagToSubcommands())
+
 	cmd.AddCommand(createCmds...)
 	cmd.AddCommand(updateCmds...)
 	cmd.AddCommand(deleteCmds...)
@@ -85,6 +97,7 @@ func NewKeysCommand() *cobra.Command {
 	cmd.AddCommand(importCmds...)
 	cmd.AddCommand(signMsgCmds...)
 	cmd.AddCommand(signTxCmds...)
+	cmd.AddCommand(slipCmds...)
 
 	return cmd
 }
@@ -351,7 +364,7 @@ func keysExportCommands() []*cobra.Command {
 
 				logger.Global.Info().Str("output_file", outputFile).Msg("Exporting private key string to file...")
 
-				return converters.WriteOutput(exportString, outputFile)
+				return utils.WriteOutput(exportString, outputFile)
 			},
 		},
 	}
@@ -372,7 +385,7 @@ func keysImportCommands() []*cobra.Command {
 				if len(args) == 1 {
 					privateKeyString = args[0]
 				} else if inputFile != "" {
-					privateKeyBz, err := converters.ReadInput(inputFile)
+					privateKeyBz, err := utils.ReadInput(inputFile)
 					privateKeyString = string(privateKeyBz)
 					if err != nil {
 						return err
@@ -566,11 +579,11 @@ func keysSignTxCommands() []*cobra.Command {
 				}
 
 				// Unmarshal Tx from input file
-				txBz, err := converters.ReadInput(inputFile)
+				txBz, err := utils.ReadInput(inputFile)
 				if err != nil {
 					return err
 				}
-				txProto := new(utilTypes.Transaction)
+				txProto := new(coreTypes.Transaction)
 				if err := codec.GetCodec().Unmarshal(txBz, txProto); err != nil {
 					return err
 				}
@@ -587,7 +600,7 @@ func keysSignTxCommands() []*cobra.Command {
 				}
 
 				// Add signature to the transaction
-				sig := new(utilTypes.Signature)
+				sig := new(coreTypes.Signature)
 				sig.PublicKey = privKey.PublicKey().Bytes()
 				sig.Signature = sigBz
 				txProto.Signature = sig
@@ -598,7 +611,7 @@ func keysSignTxCommands() []*cobra.Command {
 					return err
 				}
 
-				if err := converters.WriteOutput(txBz, outputFile); err != nil {
+				if err := utils.WriteOutput(txBz, outputFile); err != nil {
 					return err
 				}
 
@@ -638,11 +651,11 @@ func keysSignTxCommands() []*cobra.Command {
 				}
 
 				// Unmarshal Tx from input file
-				txBz, err := converters.ReadInput(inputFile)
+				txBz, err := utils.ReadInput(inputFile)
 				if err != nil {
 					return err
 				}
-				txProto := new(utilTypes.Transaction)
+				txProto := new(coreTypes.Transaction)
 				if err := codec.GetCodec().Unmarshal(txBz, txProto); err != nil {
 					return err
 				}
@@ -673,6 +686,56 @@ func keysSignTxCommands() []*cobra.Command {
 				}
 
 				logger.Global.Info().Str("address", addrHex).Bool("valid", valid).Msg("Signature checked")
+
+				return nil
+			},
+		},
+	}
+	return cmds
+}
+
+func keysSlipCommands() []*cobra.Command {
+	cmds := []*cobra.Command{
+		{
+			Use:     "DeriveChild <parentAddrHex> <index>",
+			Short:   "Derive the child key at the given index from a parent key",
+			Long:    "Derive the child key at <index> from the parent key provided optionally store it in the keybase with [--store_child]",
+			Aliases: []string{"derivechild"},
+			Args:    cobra.ExactArgs(2),
+			RunE: func(cmd *cobra.Command, args []string) error {
+				// Unpack CLI args
+				parentAddr := args[0]
+				idx64, err := strconv.ParseUint(args[1], 10, 32)
+				if err != nil {
+					return err
+				}
+				index := uint32(idx64)
+
+				// Open the debug keybase at the specified path
+				pocketDir := strings.TrimSuffix(dataDir, "/")
+				keybasePath, err := filepath.Abs(pocketDir + keybaseSuffix)
+				if err != nil {
+					return err
+				}
+				kb, err := keybase.NewKeybase(keybasePath)
+				if err != nil {
+					return err
+				}
+
+				if !nonInteractive {
+					pwd = readPassphrase(pwd)
+				}
+
+				kp, err := kb.DeriveChildFromKey(parentAddr, pwd, index, childPwd, childHint, storeChild)
+				if err != nil {
+					return err
+				}
+
+				if err := kb.Stop(); err != nil {
+					return err
+				}
+
+				logger.Global.Info().Str("address", kp.GetAddressString()).Str("parent", parentAddr).Uint32("index", index).Bool("stored", storeChild).Msg("Child key derived")
 
 				return nil
 			},