+++ consensus/e2e_tests/state_sync_test.go
@@ -225,7 +225,6 @@ func TestStateSync_UnsynchedPeerSynchs_Success(t *testing.T) {
 		Height:            testHeight,
 		StateHash:         stateHash,
 		PrevStateHash:     "",
-		NumTxs:            0,
 		ProposerAddress:   consensusPK.Address(),
 		QuorumCertificate: nil,
 	}