+++ p2p/utils_test.go
@@ -9,26 +9,29 @@ import (
 	"time"
 
 	"github.com/golang/mock/gomock"
+	"github.com/stretchr/testify/require"
+
 	"github.com/pokt-network/pocket/p2p/providers/addrbook_provider"
 	"github.com/pokt-network/pocket/p2p/providers/current_height_provider"
 	typesP2P "github.com/pokt-network/pocket/p2p/types"
 	mocksP2P "github.com/pokt-network/pocket/p2p/types/mocks"
 	"github.com/pokt-network/pocket/runtime"
 	"github.com/pokt-network/pocket/runtime/configs"
-	types "github.com/pokt-network/pocket/runtime/configs/types"
+	"github.com/pokt-network/pocket/runtime/configs/types"
+	"github.com/pokt-network/pocket/runtime/defaults"
 	"github.com/pokt-network/pocket/runtime/genesis"
+	"github.com/pokt-network/pocket/runtime/test_artifacts"
 	coreTypes "github.com/pokt-network/pocket/shared/core/types"
 	cryptoPocket "github.com/pokt-network/pocket/shared/crypto"
 	"github.com/pokt-network/pocket/shared/modules"
 	mockModules "github.com/pokt-network/pocket/shared/modules/mocks"
 	"github.com/pokt-network/pocket/telemetry"
-	"github.com/stretchr/testify/require"
 )
 
 // ~~~~~~ RainTree Unit Test Configurations ~~~~~~
 
 const (
-	serviceUrlFormat  = "node%d.consensus:8080"
+	serviceURLFormat  = "node%d.consensus:42069"
 	eventsChannelSize = 10000
 	// Since we simulate up to a 27 node network, we will pre-generate a n >= 27 number of keys to avoid generation
 	// every time. The genesis config seed start is set for deterministic key generation and 42 was chosen arbitrarily.
@@ -70,7 +73,7 @@ type TestNetworkSimulationConfig map[string]struct {
 
 // CLEANUP: This could (should?) be a codebase-wide shared test helper
 func validatorId(i int) string {
-	return fmt.Sprintf(serviceUrlFormat, i)
+	return fmt.Sprintf(serviceURLFormat, i)
 }
 
 func waitForNetworkSimulationCompletion(t *testing.T, wg *sync.WaitGroup) {
@@ -118,7 +121,7 @@ func createMockRuntimeMgrs(t *testing.T, numValidators int) []modules.RuntimeMgr
 			PrivateKey:    valKeys[i].String(),
 			P2P: &configs.P2PConfig{
 				PrivateKey:     valKeys[i].String(),
-				ConsensusPort:  8080,
+				Port:           defaults.DefaultP2PPort,
 				UseRainTree:    true,
 				ConnectionType: types.ConnectionType_EmptyConnection,
 			},
@@ -165,8 +168,8 @@ func createMockGenesisState(valKeys []cryptoPocket.PrivateKey) *genesis.GenesisS
 			ActorType:       coreTypes.ActorType_ACTOR_TYPE_VAL,
 			Address:         addr,
 			PublicKey:       valKey.PublicKey().String(),
-			GenericParam:    validatorId(i + 1),
-			StakedAmount:    "1000000000000000",
+			ServiceUrl:      validatorId(i + 1),
+			StakedAmount:    test_artifacts.DefaultStakeAmountString,
 			PausedHeight:    int64(0),
 			UnstakingHeight: int64(0),
 			Output:          addr,