+++ persistence/test/generic_test.go
@@ -12,8 +12,6 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-// TODO(andrew): Be consistent with `GenericParam` and `ActorSpecificParam` throughout the codebase; preferably the latter.
-
 func getGenericActor[T any](
 	protocolActorSchema types.ProtocolActorSchema,
 	getActor func(*persistence.PostgresContext, []byte) (T, error),
@@ -214,7 +212,7 @@ func getActorValues(_ types.ProtocolActorSchema, actorValue reflect.Value) *core
 		Address:         actorValue.FieldByName("Address").String(),
 		PublicKey:       actorValue.FieldByName("PublicKey").String(),
 		StakedAmount:    actorValue.FieldByName("StakedAmount").String(),
-		GenericParam:    actorValue.FieldByName("GenericParam").String(),
+		ServiceUrl:      actorValue.FieldByName("ServiceUrl").String(),
 		Output:          actorValue.FieldByName("Output").String(),
 		PausedHeight:    actorValue.FieldByName("PausedHeight").Int(),
 		UnstakingHeight: actorValue.FieldByName("UnstakingHeight").Int(),