+++ runtime/defaults/defaults.go
@@ -18,6 +18,7 @@ const (
 
 var (
 	DefaultRemoteCLIURL = fmt.Sprintf("http://%s:%s", DefaultRPCHost, DefaultRPCPort)
+	DefaultUseLibp2p    = false
 
 	// consensus
 	DefaultConsensusMaxMempoolBytes = uint64(500000000)
@@ -29,10 +30,10 @@ var (
 	DefaultUtilityMaxMempoolTransactionBytes = uint64(1024 ^ 3) // 1GB V0 defaults
 	DefaultUtilityMaxMempoolTransactions     = uint32(9000)
 	// persistence
-	DefaultPersistencePostgresUrl    = "postgres://postgres:postgres@pocket-db:5432/postgres"
+	DefaultPersistencePostgresURL    = "postgres://postgres:postgres@pocket-db:5432/postgres"
 	DefaultPersistenceBlockStorePath = "/var/blockstore"
 	// p2p
-	DefaultP2PConsensusPort   = uint32(8080)
+	DefaultP2PPort            = uint32(42069)
 	DefaultP2PUseRainTree     = true
 	DefaultP2PConnectionType  = types.ConnectionType_TCPConnection
 	DefaultP2PMaxMempoolCount = uint64(1e5)