+++ runtime/test_artifacts/defaults.go
@@ -3,21 +3,19 @@ package test_artifacts
 import (
 	"math/big"
 
-	"github.com/pokt-network/pocket/shared/converters"
+	"github.com/pokt-network/pocket/shared/utils"
 )
 
 var (
 	DefaultChains              = []string{"0001"}
 	DefaultServiceURL          = ""
 	DefaultStakeAmount         = big.NewInt(1000000000000)
-	DefaultStakeAmountString   = converters.BigIntToString(DefaultStakeAmount)
-	DefaultMaxRelays           = big.NewInt(1000000)
-	DefaultMaxRelaysString     = converters.BigIntToString(DefaultMaxRelays)
+	DefaultStakeAmountString   = utils.BigIntToString(DefaultStakeAmount)
 	DefaultAccountAmount       = big.NewInt(100000000000000)
-	DefaultAccountAmountString = converters.BigIntToString(DefaultAccountAmount)
+	DefaultAccountAmountString = utils.BigIntToString(DefaultAccountAmount)
 	DefaultPauseHeight         = int64(-1)
 	DefaultUnstakingHeight     = int64(-1)
 	DefaultChainID             = "testnet"
-	ServiceUrlFormat           = "node%d.consensus:8080"
+	ServiceURLFormat           = "node%d.consensus:42069"
 	DefaultMaxBlockBytes       = uint64(4000000)
 )