+++ logger/docs/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.0.9] - 2023-02-28
+
+- Removed the unused `bus` from the `logger` struct
+
 ## [0.0.0.8] - 2023-02-24
 
 - Update the logger module interface to use logger pointers instead of values